Completed code immediateorder. Removed conditions by keeping masked value. Refactor simpletracking generator to be used by `listing`.
