+++ babisteps/basemodels/immediateorder.py
@@ -5,8 +5,7 @@ from typing import Any, Callable, Literal, Optional, get_type_hints
 import networkx as nx
 import numpy as np
 from pydantic import BaseModel, model_validator
-from sparse._dok import DOK
-from sparse._sparse_array import SparseArray
+from sparse import DOK, SparseArray
 
 from babisteps.basemodels.FOL import FOL, Exists, IsRelated
 from babisteps.basemodels.generators import BaseGenerator
@@ -21,8 +20,8 @@ class ImmediateOrderRequest(BaseModel):
     e1: Optional[Any] = None
     # Relation
     r: Optional[Any] = None
-    relation_type: Literal['relative_event', 'relative_size',
-                           'relative_position', 'absolute_position']
+    relation_type: Literal["relative_event", "relative_size",
+                           "relative_position", "absolute_position"]
     shape_str: Literal[("locations", ), ("actors", ), ("objects", ),
                        ("events", )]
 
@@ -37,7 +36,6 @@ class ImmediateOrderRequestPolar(ImmediateOrderRequest):
     answer: Literal["yes", "no", "unknown"]
 
     def get_question(self):
-
         if self.shape_str in [("locations", ), ("objects", )]:
             return (f"Is the {self.e0.name} {random.choice(self.r.base)} "
                     f"the {self.e1.name}?")
@@ -51,43 +49,101 @@ class ImmediateOrderRequestPolar(ImmediateOrderRequest):
                 "Invalid shape_str for ImmediateOrderRequestPolar")
 
     def get_answer(self):
-        return self.answer
+        return [self.answer]
 
 
 class ImmediateOrderRequestHow(ImmediateOrderRequest):
     answer: Literal["designated_relation", "unknown"]
 
     def get_question(self):
-        return f"How is {self.e1.name} related to {self.e0.name}?"
+        if self.shape_str in [("locations", ), ("objects", )]:
+            return f"How is the {self.e1.name} related to the {self.e0.name}?"
+        elif self.shape_str == ("actors", ):
+            return f"How is {self.e1.name} related to {self.e0.name}?"
+        elif self.shape_str == ("events", ):
+            return f"How was the {self.e1.name} related to the {self.e0.name}?"
+        else:
+            raise ValueError("Invalid shape_str for ImmediateOrderRequestHow")
 
-    def get_answer(self):
-        # TODO: Probably the way to get the answer should be different
-        # and depend on the relation type?
+    def _get_answer_from_relation(self, relation: Relationship):
         if self.answer == "designated_relation":
-            return self.r0
+            answers = []
+            if self.shape_str in [("locations", ), ("objects", )]:
+                for i in relation.base:
+                    answers.append(
+                        f"The {self.e0.name} is {i} the {self.e1.name}")
+                for j in relation.opposite:
+                    answers.append(
+                        f"The {self.e1.name} is {j} the {self.e0.name}")
+
+            elif self.shape_str == ("actors", ):
+                for i in relation.base:
+                    answers.append(f"{self.e0.name} is {i} {self.e1.name}")
+                for j in relation.opposite:
+                    answers.append(f"{self.e1.name} is {j} {self.e0.name}")
+
+            elif self.shape_str == ("events", ):
+                for i in relation.base:
+                    answers.append(
+                        f"The {self.e0.name} was {i} the {self.e1.name}")
+                for j in relation.opposite:
+                    answers.append(
+                        f"The {self.e1.name} was {j} the {self.e0.name}")
+            else:
+                raise ValueError(
+                    "Invalid shape_str for ImmediateOrderRequestHow")
+            return answers
+
         elif self.answer == "unknown":
-            return self.answer
+            return [self.answer]
         else:
             raise ValueError(
                 "'answer' must be 'designated_relation' or 'unknown'")
 
+    def get_answer(self):
+        try:
+            answer = self._get_answer_from_relation(self.r)
+            return answer
+        except Exception as e:
+            raise e
+
+    def get_options(self, relations: list[Relationship]):
+        """
+        Extracts and returns a shuffled list of options from the provided relationships.
+        """
+        # keep current value of answer to then return it to its original value
+        original_answer = self.answer
+        # set answer to 'designated_relation' to produce all the options
+        self.answer = "designated_relation"
+        options = []
+        for relation in relations:
+            options.extend(self._get_answer_from_relation(relation))
+        # now set the answer back to its original value
+        self.answer = original_answer
+        return options
+
 
 class ImmediateOrderRequestWhat(ImmediateOrderRequest):
-    answer: Literal["second_designated_event", "unknown"]
+    answer: Literal["second_entity", "none", "unknown"]
 
     def get_question(self):
-        # TODO: The question generation is related to the relation type
-        # probably this would need to be handle differently.
-        return f"To what is {self.e0.name} {self.r0}?"
+        if self.shape_str in [("locations", ), ("objects", )]:
+            return f"To what is the {self.e0.name} {random.choice(self.r.base)}?"
+        elif self.shape_str == ("actors", ):
+            return f"To who is {self.e0.name} {random.choice(self.r.base)}?"
+        elif self.shape_str == ("events", ):
+            return f"To what was the {self.e0.name} {random.choice(self.r.base)}?"
+        else:
+            raise ValueError("Invalid shape_str for ImmediateOrderRequestWhat")
 
     def get_answer(self):
-        if self.answer == "second_designated_event":
-            return self.e1.name
-        elif self.answer == "unknown":
-            return self.answer
+        if self.answer == "second_entity":
+            return [self.e1.name]
+        elif self.answer == "none" or self.answer == "unknown":
+            return [self.answer]
         else:
             raise ValueError(
-                "'answer' must be 'second_designated_event' or 'unknown'")
+                "'answer' must be 'second_entity', 'none', or 'unknown'")
 
 
 class ImmediateOrderModel(BaseModel):
@@ -129,6 +185,7 @@ class ImmediateOrder(BaseGenerator):
         loader_mapping: dict[type[ImmediateOrderRequest], Callable] = {
             ImmediateOrderRequestPolar: self._immediate_order_polar,
             ImmediateOrderRequestHow: self._immediate_order_how,
+            ImmediateOrderRequestWhat: self._immediate_order_what,
         }
         # Get the type of the answer
         topic_type = type(self.topic)
@@ -149,25 +206,29 @@ class ImmediateOrder(BaseGenerator):
                 g.add_edge(i, j)
         return matrix, g
 
-    def _fill_edges(self,
-                    matrix: SparseArray,
-                    g: nx.DiGraph,
-                    n: int,
-                    condition: Optional[Callable] = None):
-        nans = np.argwhere(np.isnan(matrix.todense()))
+    def _fill_edges(
+        self,
+        matrix: SparseArray,
+        g: nx.DiGraph,
+        n: int,
+    ):
+        nans = np.isnan(matrix.todense())
+        origin_mask = ~nans
+        origin_vals = matrix.todense()[origin_mask]
+        nans = np.argwhere(nans)
         while len(nans) > 0:
             c = np.random.choice(len(nans))
             c = nans[c]
             i, j = int(c[0]), int(c[1])
-
             matrix_aux = deepcopy(matrix)
             g_aux = deepcopy(g)
             matrix_aux[i, j] = 1
             matrix_aux[j, i] = 0
             g_aux.add_edge(i, j)
             matrix_aux, g_aux = self._transitive_closure(matrix_aux, g_aux)
-            # if condition is not none, check it
-            if condition and not condition(matrix_aux, g_aux):
+            # validate that origin_vals are not changed
+            new_vals = matrix_aux.todense()[origin_mask]
+            if not np.array_equal(origin_vals, new_vals):
                 matrix[i, j] = 0
                 continue
             if len(g_aux.edges) == n:
@@ -202,7 +263,6 @@ class ImmediateOrder(BaseGenerator):
         self.topic.e1 = e1
         self.topic.r = self.model.relations[0]
         r_am, r = self._create_empty_graph()
-        condition = None  # Re-defined only in the unknown case
         if self.topic.answer == "yes":
             r_am[0, 1] = 1
             r_am[1, 0] = 0
@@ -214,18 +274,15 @@ class ImmediateOrder(BaseGenerator):
         elif self.topic.answer == "unknown":
             r_am[0, 1] = 0
             r_am[1, 0] = 0
-            # validate that both 0,1 and 1,0 are not in the graph
-            #condition = lambda r_am, r: (r_am[0, 1] == 0 and r_am[1, 0] == 0) and /
-            # (not r.has_edge(0, 1) and not r.has_edge(1, 0))
-            condition = lambda r_am, r: not r.has_edge(
-                0, 1) and not r.has_edge(1, 0)
 
-        self.logger.debug("Creating Immediate Order Graph",
-                          answer=self.topic.answer,
-                          e0=e0.name,
-                          e1=e1.name,
-                          relation=self.topic.r.name)
-        r_am, r = self._fill_edges(r_am, r, self.edge_qty, condition)
+        self.logger.debug(
+            "Creating Immediate Order Graph",
+            answer=self.topic.answer,
+            e0=e0.name,
+            e1=e1.name,
+            relation=self.topic.r.name,
+        )
+        r_am, r = self._fill_edges(r_am, r, self.edge_qty)
         graphs.append(
             ImmediateGraph(am=r_am,
                            g=r,
@@ -246,6 +303,7 @@ class ImmediateOrder(BaseGenerator):
         self._create_aux()
         self.topic.e0 = e0
         self.topic.e1 = e1
+        self.topic.r = self.model.relations[0]
         r_am, r = self._create_empty_graph()
         if self.topic.answer == "designated_relation":
             r_am[0, 1] = 1
@@ -266,10 +324,57 @@ class ImmediateOrder(BaseGenerator):
         if len(self.model.relations) > 1:
             for i, rlt in enumerate(self.model.relations[1:], start=1):
                 g_am, g = self._create_empty_graph()
-                if (self.topic.answer == "designated_relation"
-                        or self.topic.answer == "unknown"):
-                    g_am[0, 1] = 0
-                    g_am[1, 0] = 0
+                g_am[0, 1], g_am[1, 0] = 0, 0
+                g_am, g = self._fill_edges(g_am, g, self.edge_qty)
+                graphs.append(
+                    ImmediateGraph(am=g_am, g=g, name=rlt.name, index=i))
+        return graphs
+
+    def _immediate_order_what(self):
+        e0 = self.model.entities[0]
+        e1 = self.model.entities[1]
+        graphs = []
+        self._create_aux()
+        self.topic.e0 = e0
+        self.topic.e1 = e1
+        self.topic.r = self.model.relations[0]
+        r_am, r = self._create_empty_graph()
+        if self.topic.answer == "second_entity":
+            r_am[0, 1] = 1
+            r_am[1, 0] = 0
+            r.add_edge(0, 1)
+            r_am[0, 2:] = 0
+        elif self.topic.answer == "none":
+            if self.edge_qty <= len(self.model.entities) - 1:
+                self.logger.error(
+                    "The #edges must be <= that # entities when the answer is 'none'",
+                    answer=self.topic.answer,
+                    edge_qty=self.edge_qty,
+                    n_entities=len(self.model.entities),
+                )
+                raise ValueError(
+                    "The #edges must be <= that # entities when the answer is 'none'"
+                )
+            r_am[0, :] = 0
+            r_am[1:, 0] = 1
+            for i in range(1, len(self.model.entities)):
+                r.add_edge(i, 0)
+        elif self.topic.answer == "unknown":
+            rnd_e = random.choice(range(1, len(self.model.entities)))
+            r_am[rnd_e, 0] = 0
+            r_am[0, 1:] = 0
+        else:
+            raise ValueError(
+                "Invalid answer should be 'second_entity' or 'unknown'")
+        r_am, r = self._fill_edges(r_am, r, self.edge_qty)
+        graphs.append(
+            ImmediateGraph(am=r_am,
+                           g=r,
+                           name=self.model.relations[0].name,
+                           index=0))
+        if len(self.model.relations) > 1:
+            for i, rlt in enumerate(self.model.relations[1:], start=1):
+                g_am, g = self._create_empty_graph()
                 g_am, g = self._fill_edges(g_am, g, self.edge_qty)
                 graphs.append(
                     ImmediateGraph(am=g_am, g=g, name=rlt.name, index=i))
@@ -285,26 +390,28 @@ class ImmediateOrder(BaseGenerator):
 
         def describe_relation(relation, graph, shape_str):
             graph_sentences = []
-            for (i, j) in graph.am.data:
+            for i, j in graph.am.data:
                 # check first am
                 edge = graph.am[i, j]
                 if edge == 1:
                     # then verify if the edge exists in the graph
                     assert graph.g.has_edge(
-                        i, j), "edge {}-{} does not exist in the graph".format(
-                            i, j)
+                        i,
+                        j), ("edge {}-{} does not exist in the graph".format(
+                            i, j))
                     graph_sentences.append(
-                        IsRelated(relation=relation,
-                                  entity0=self.model.entities[i],
-                                  entity1=self.model.entities[j],
-                                  shape_str=shape_str))
+                        IsRelated(
+                            relation=relation,
+                            entity0=self.model.entities[i],
+                            entity1=self.model.entities[j],
+                            shape_str=shape_str,
+                        ))
             return graph_sentences
 
         world_enumerate = []
         story = []
 
         for t, dim_str in zip(self.model.as_tuple, self.shape_str):
-
             world_enumerate.extend(enumerate_model(t, dim_str))
         random.shuffle(world_enumerate)
 
@@ -331,23 +438,23 @@ class ImmediateOrder(BaseGenerator):
 
     def get_json(self):
         json = self.story.create_json()
-        options = list(get_type_hints(self.topic)['answer'].__args__)
+        options = list(get_type_hints(self.topic)["answer"].__args__)
         if isinstance(self.topic, ImmediateOrderRequestPolar):
-            # do nothing
             pass
         elif isinstance(self.topic, ImmediateOrderRequestHow):
-            options.remove('designated_relation')
-            options.extend([r.name for r in self.model.relations])
+            options.remove("designated_relation")
+            o = self.topic.get_options(self.model.relations)
+            options.extend(o)
         elif isinstance(self.topic, ImmediateOrderRequestWhat):
-            options.remove('second_designated_event')
+            options.remove("second_entity")
             options.extend([e.name for e in self.model.entities])
 
         random.shuffle(options)
-        json['options'] = options
+        json["options"] = options
         if self.name:
-            json['leaf'] = self.name.split('_-_')[0]
-            json['leaf_label'] = self.name.split('_-_')[1]
-            json['leaf_index'] = self.name.split('_-_')[2]
+            json["leaf"] = self.name.split("_-_")[0]
+            json["leaf_label"] = self.name.split("_-_")[1]
+            json["leaf_index"] = self.name.split("_-_")[2]
         return json
 
     def get_txt(self):
