+++ babisteps/tasks/immediateorder/config.yaml
@@ -1,5 +1,5 @@
 task_name: immediateorder
 relations_qty: 3
-entities: 8
+entities: 7
 gen_kwargs:
 func: utils._get_generators
